修复了查找时不会查找iMemTable，可能会导致查找不到已添加数据的bug
