+++ Database.go
@@ -6,7 +6,6 @@ import (
 	"log"
 	"os"
 	"path"
-	"sync"
 )
 
 type Database struct {
@@ -16,23 +15,6 @@ type Database struct {
 	iMemTable *ReadOnlyMemTables
 	// SSTable 列表
 	TableTree *ssTable.TableTree
-}
-
-type ReadOnlyMemTables struct {
-	// 只读内存表
-	readonlyTable []*MemTable
-	lock          sync.Locker
-}
-
-func (r *ReadOnlyMemTables) Init() {
-	r.readonlyTable = make([]*MemTable, 0)
-	r.lock = &sync.Mutex{}
-}
-
-func (r *ReadOnlyMemTables) Getlen() int {
-	r.lock.Lock()
-	r.lock.Unlock()
-	return len(r.readonlyTable)
 }
 
 // 数据库，全局唯一实例
@@ -56,29 +38,15 @@ func (d *Database) loadAllWalFiles(dir string) {
 				Wal:        preWal,
 			}
 			log.Printf("add table to iMemTable, table: %v\n", table)
-			d.iMemTable.Add(table)
+			d.iMemTable.AddTable(table)
 		}
 	}
 	return
-}
-
-func (r *ReadOnlyMemTables) Add(table *MemTable) {
-	r.lock.Lock()
-	r.readonlyTable = append(r.readonlyTable, table)
-	r.lock.Unlock()
-}
-
-func (r *ReadOnlyMemTables) Get() *MemTable {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-	table := r.readonlyTable[0]
-	r.readonlyTable = r.readonlyTable[1:]
-	return table
 }
 
 func (d *Database) Swap() {
 	table := d.MemTable.Swap()
 	// 将内存表存储到 iMemTable 中
 	log.Printf("add table to iMemTable, table: %v\n", table)
-	d.iMemTable.Add(table)
+	d.iMemTable.AddTable(table)
 }
